WIP of SPMW
SPMW - set protected memory window
The problem I'm realizing is that now load() needs to
be concerned not only with the current cpu_state, but
also the _next_ cpu_state. It needs to know the current
state so it can allow or deny access to the protected
region as well as knowing what the protected region
even is. It then needs to set the next cpu_state
because if an invalid access was made, it should set
the exception flag and set the IP to the the exception
vector address (EVA).

I think the way to do this will be to wrap every
instruction in it's own function that has more or less
the same signature as cpu() - takes a load(), a store(),
the current cpu_state and returns a new cpu_state.